refactor(agent): add presenter lifecycle
