+++ src/main/env.d.ts
@@ -5,6 +5,8 @@ interface ImportMetaEnv {
   readonly VITE_GITHUB_CLIENT_SECRET: string
   readonly VITE_GITHUB_REDIRECT_URI: string
   readonly VITE_LOG_IPC_CALL: string
+  readonly VITE_AGENT_PRESENTER_ENABLED?: string
+  readonly VITE_AGENT_PRESENTER_DEBUG?: string
   readonly VITE_APP_LIFECYCLE_HOOK_DELAY?: string
   readonly VITE_PROVIDER_DB_URL?: string
 }
