+++ src/main/presenter/agentPresenter/tool/index.ts
@@ -1,0 +1,6 @@
+export { ToolCallCenter } from './toolCallCenter'
+export { ToolRegistry } from './toolRegistry'
+export { resolveToolRoute } from './toolRouter'
+export type { ToolCallContext } from './toolCallCenter'
+export type { ToolRegistryEntry, ToolRegistrySource } from './toolRegistry'
+export type { ToolRouteDecision } from './toolRouter'
