+++ src/main/presenter/agentPresenter/tool/toolRouter.ts
@@ -1,0 +1,10 @@
+import type { ToolRegistrySource } from './toolRegistry'
+
+export type ToolRouteDecision = {
+  target: ToolRegistrySource
+  permissionType?: 'read' | 'write' | 'all' | 'command'
+}
+
+export function resolveToolRoute(_toolName: string): ToolRouteDecision {
+  return { target: 'mcp' }
+}
