+++ src/main/presenter/index.ts
@@ -21,6 +21,7 @@ import {
   ISyncPresenter,
   ITabPresenter,
   IThreadPresenter,
+  IAgentPresenter,
   IUpgradePresenter,
   IWindowPresenter,
   IWorkspacePresenter,
@@ -47,6 +48,7 @@ import { KnowledgePresenter } from './knowledgePresenter'
 import { WorkspacePresenter } from './workspacePresenter'
 import { ToolPresenter } from './toolPresenter'
 import { CommandPermissionHandler } from './threadPresenter/handlers/commandPermissionHandler'
+import { AgentPresenter } from './agentPresenter'
 
 // IPC调用上下文接口
 interface IPCCallContext {
@@ -71,6 +73,7 @@ export class Presenter implements IPresenter {
   llmproviderPresenter: ILlmProviderPresenter
   configPresenter: IConfigPresenter
   threadPresenter: IThreadPresenter
+  agentPresenter: IAgentPresenter
   devicePresenter: IDevicePresenter
   upgradePresenter: IUpgradePresenter
   shortcutPresenter: IShortcutPresenter
@@ -111,6 +114,10 @@ export class Presenter implements IPresenter {
       this.configPresenter,
       commandPermissionHandler
     )
+    this.agentPresenter = new AgentPresenter({
+      threadPresenter: this.threadPresenter,
+      configPresenter: this.configPresenter
+    })
     this.mcpPresenter = new McpPresenter(this.configPresenter)
     this.upgradePresenter = new UpgradePresenter(this.configPresenter)
     this.shortcutPresenter = new ShortcutPresenter(this.configPresenter)
@@ -301,9 +308,34 @@ ipcMain.handle(
         )
       }
 
+      const useAgentPresenter = import.meta.env.VITE_AGENT_PRESENTER_ENABLED === '1'
+
       // 通过名称获取对应的 Presenter 实例
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const calledPresenter: any = presenter[name as keyof Presenter]
+      let calledPresenter: any = presenter[name as keyof Presenter]
+      let resolvedMethod = method
+      let resolvedPayloads = payloads
+
+      if (useAgentPresenter && name === 'threadPresenter' && presenter.agentPresenter) {
+        if (method === 'sendMessage') {
+          const [conversationId, content] = payloads as [string, string]
+          calledPresenter = presenter.agentPresenter
+          resolvedMethod = 'sendMessage'
+          resolvedPayloads = [conversationId, content, tabId]
+        } else if (method === 'continueStreamCompletion') {
+          calledPresenter = presenter.agentPresenter
+          resolvedMethod = 'continueLoop'
+          resolvedPayloads = payloads
+        } else if (method === 'stopMessageGeneration') {
+          calledPresenter = presenter.agentPresenter
+          resolvedMethod = 'cancelLoop'
+          resolvedPayloads = payloads
+        } else if (method === 'handlePermissionResponse') {
+          calledPresenter = presenter.agentPresenter
+          resolvedMethod = 'handlePermissionResponse'
+          resolvedPayloads = payloads
+        }
+      }
 
       if (!calledPresenter) {
         console.warn(`[IPC Warning] Tab:${context.tabId} calling wrong presenter: ${name}`)
@@ -311,9 +343,9 @@ ipcMain.handle(
       }
 
       // 检查方法是否存在且为函数
-      if (isFunction(calledPresenter, method)) {
+      if (isFunction(calledPresenter, resolvedMethod)) {
         // 调用方法并返回结果
-        return calledPresenter[method](...payloads)
+        return calledPresenter[resolvedMethod](...resolvedPayloads)
       } else {
         console.warn(
           `[IPC Warning] Tab:${context.tabId} called method is not a function or does not exist: ${name}.${method}`
