+++ src/shared/types/index.d.ts
@@ -1,5 +1,7 @@
 // Temporary barrel: keep legacy presenters to avoid breaking changes during migration
 export type * from './presenters/legacy.presenters'
+export type { IAgentPresenter } from './presenters/agent.presenter'
+export type { MESSAGE } from './presenters/thread.presenter'
 export type * from './presenters/agent-provider'
 export type * from './presenters/workspace'
 export type * from './presenters/tool.presenter'
