+++ src/shared/types/presenters/index.d.ts
@@ -49,5 +49,8 @@ export type {
 // Tool Presenter types
 export type { IToolPresenter } from './tool.presenter'
 
+// Agent Presenter types
+export type { IAgentPresenter } from './agent.presenter'
+
 // Re-export legacy types temporarily for compatibility
 export * from './legacy.presenters'
