+++ test/main/presenter/agentPresenter/agentPresenter.test.ts
@@ -1,0 +1,8 @@
+import { describe, it, expect } from 'vitest'
+import * as agentPresenterModule from '../../../../src/main/presenter/agentPresenter'
+
+describe('AgentPresenter scaffolding', () => {
+  it('loads module entrypoint', () => {
+    expect(agentPresenterModule).toBeDefined()
+  })
+})
